move getSingleInstance
